+++ lib/segment/src/index/field_index/numeric_index/builders.rs
@@ -7,11 +7,10 @@ use common::types::PointOffsetType;
 use gridstore::Blob;
 use serde_json::Value;
 
-use super::Encodable;
-use super::index::{NumericIndex, NumericIndexIntoInnerValue};
 use super::mmap_numeric_index::UniversalNumericIndex;
 use super::mutable_numeric_index::InMemoryNumericIndex;
 use super::storage::NumericIndexInner;
+use super::{Encodable, NumericIndex, NumericIndexIntoInnerValue};
 use crate::common::operation_error::{OperationError, OperationResult};
 use crate::index::field_index::numeric_point::Numericable;
 use crate::index::field_index::stored_point_to_values::StoredValue;
