+++ lib/segment/src/index/field_index/numeric_index/mod.rs
@@ -1,6 +1,6 @@
 mod builders;
+mod encodable;
 pub mod immutable_numeric_index;
-mod index;
 mod lifecycle;
 pub mod mmap_numeric_index;
 pub mod mutable_numeric_index;
@@ -10,14 +10,31 @@ mod read_ops;
 mod storage;
 mod value_indexer;
 
+use std::marker::PhantomData;
+
 pub use builders::{NumericIndexBuilder, NumericIndexGridstoreBuilder, NumericIndexMmapBuilder};
-pub use index::{NumericIndex, NumericIndexIntoInnerValue};
-pub use lifecycle::Encodable;
+pub use encodable::Encodable;
+use gridstore::Blob;
 pub use numeric_field_index::{NumericFieldIndex, NumericFieldIndexRead};
 pub use numeric_index_read::NumericIndexRead;
 pub use read_ops::StreamRange;
 pub use storage::NumericIndexInner;
 pub use storage::read_only::ReadOnlyNumericIndexInner;
 
+use crate::index::field_index::numeric_point::Numericable;
+use crate::index::field_index::stored_point_to_values::StoredValue;
+
 #[cfg(test)]
 mod tests;
+
+pub struct NumericIndex<T: Encodable + Numericable + StoredValue + Send + Sync + Default, P>
+where
+    Vec<T>: Blob,
+{
+    inner: NumericIndexInner<T>,
+    _phantom: PhantomData<P>,
+}
+
+pub trait NumericIndexIntoInnerValue<T, P> {
+    fn into_inner_value(value: P) -> T;
+}
